Added sheet for the update details
