+++ src/pages/Encounters/tabs/overview/summary-panel-details-tab.tsx
@@ -1,31 +1,4 @@
-import { useQuery } from "@tanstack/react-query";
-import { format } from "date-fns";
-import { Signal, SquarePen } from "lucide-react";
-import { Link } from "raviger";
-import React from "react";
-import { useTranslation } from "react-i18next";
-
-import CareIcon from "@/CAREUI/icons/CareIcon";
-
-import { Badge } from "@/components/ui/badge";
-import { Button } from "@/components/ui/button";
-import { Separator } from "@/components/ui/separator";
-
-import { Avatar } from "@/components/Common/Avatar";
-
-import query from "@/Utils/request/query";
-import { StatusBadge } from "@/pages/Encounters/EncounterProperties";
 import { useEncounter } from "@/pages/Encounters/utils/EncounterProvider";
-import {
-  AccountBillingStatus,
-  AccountStatus,
-} from "@/types/billing/account/Account";
-import accountApi from "@/types/billing/account/accountApi";
-import {
-  ENCOUNTER_CLASSES_COLORS,
-  ENCOUNTER_CLASS_ICONS,
-  ENCOUNTER_PRIORITY_COLORS,
-} from "@/types/emr/encounter/encounter";
 
 import { Account } from "./summary-panel-details-tab/account";
 import { AuditLogs } from "./summary-panel-details-tab/auditlogs";
@@ -37,289 +10,15 @@ import { Questionnaires } from "./summary-panel-details-tab/forms";
 import { HospitalizationDetails } from "./summary-panel-details-tab/hospitalisation";
 import { Locations } from "./summary-panel-details-tab/locations";
 import { ManageCareTeam } from "./summary-panel-details-tab/manage-care-team";
+import { SummaryPanelEncounterDetails } from "./summary-panel-details-tab/summary-panel-encounter-details";
 
 export const SummaryPanelDetailTab = () => {
-  const { t } = useTranslation();
-  const {
-    selectedEncounter: encounter,
-    canWriteSelectedEncounter: canEdit,
-    patientId,
-    facilityId,
-    patient,
-  } = useEncounter();
-
-  const { data: account } = useQuery({
-    queryKey: ["accounts", patientId],
-    queryFn: query(accountApi.listAccount, {
-      pathParams: { facilityId: facilityId || "" },
-      queryParams: {
-        patient: patientId,
-        status: AccountStatus.active,
-        billing_status: AccountBillingStatus.open,
-        limit: 1,
-      },
-    }),
-    enabled: !!facilityId,
-  });
-
-  if (!encounter) return null;
-  const EncounterClassIcon = ENCOUNTER_CLASS_ICONS[encounter.encounter_class];
+  const { canWriteSelectedEncounter: canEdit } = useEncounter();
 
   return (
     <div>
-      <div className="flex flex-col gap-2">
-        <div className="hidden @xs:flex flex-col sm:flex-row p-3 bg-white -mt-1 rounded-lg gap-4 shadow">
-          <div className="flex flex-col gap-4 sm:border-r border-gray-200 pr-4">
-            <div className="flex flex-row gap-14">
-              <div className="flex flex-col gap-4">
-                <div>
-                  <span className="text-sm font-medium text-gray-700">
-                    {t("status")}:
-                  </span>
-                  <div>
-                    <StatusBadge encounter={encounter} />
-                  </div>
-                </div>
-
-                <div>
-                  <span className="text-sm font-medium text-gray-700">
-                    {t("encounter_class")}:
-                  </span>
-                  <div>
-                    <Badge
-                      variant={
-                        ENCOUNTER_CLASSES_COLORS[encounter.encounter_class]
-                      }
-                    >
-                      <EncounterClassIcon className="size-3" />
-                      <span className="whitespace-nowrap">
-                        {t(`encounter_class__${encounter.encounter_class}`)}
-                      </span>
-                    </Badge>
-                  </div>
-                </div>
-                <div>
-                  <span className="text-sm font-medium text-gray-700">
-                    {t("priority")}:
-                  </span>
-                  <div>
-                    <Badge
-                      variant={ENCOUNTER_PRIORITY_COLORS[encounter.priority]}
-                    >
-                      <Signal className="size-3" />
-                      {t(`encounter_priority__${encounter.priority}`)}
-                    </Badge>
-                  </div>
-                </div>
-
-                <div>
-                  <span className="text-sm font-medium text-gray-700">
-                    {t("location")}:
-                  </span>
-                  <div>
-                    {encounter.current_location?.name ? (
-                      <Badge
-                        variant="secondary"
-                        className="inline-flex items-center gap-1.5"
-                      >
-                        <CareIcon icon="l-location-point" className="size-3" />
-                        {encounter.current_location?.name || t("none")}
-                      </Badge>
-                    ) : (
-                      <span>--</span>
-                    )}
-                  </div>
-                </div>
-              </div>
-
-              <div className="flex flex-col gap-4">
-                <div>
-                  <span className="text-sm font-medium text-gray-700">
-                    {t("dep_and_teams")}:
-                  </span>
-                  <div className="flex flex-wrap gap-2">
-                    {encounter.organizations.length > 0 ? (
-                      <>
-                        {encounter.organizations.map((org) => (
-                          <Badge
-                            key={org.id}
-                            variant="blue"
-                            className="capitalize"
-                          >
-                            {org.name}
-                          </Badge>
-                        ))}
-                      </>
-                    ) : (
-                      <span>--</span>
-                    )}
-                  </div>
-                </div>
-                <div>
-                  <span className="text-sm font-medium text-gray-700">
-                    {t("encounter_tags")}:
-                  </span>
-                  <div className="flex flex-wrap gap-2">
-                    {encounter.tags.length > 0 ? (
-                      <>
-                        {encounter.tags.map((tag) => (
-                          <Badge
-                            key={tag.id}
-                            variant="secondary"
-                            className="capitalize"
-                            title={tag.description}
-                          >
-                            {tag.display}
-                          </Badge>
-                        ))}
-                      </>
-                    ) : (
-                      <p className="text-sm text-gray-500">{t("no_tags")}</p>
-                    )}
-                  </div>
-                </div>
-
-                <div>
-                  <span className="text-sm font-medium text-gray-700">
-                    {t("hospitalisation")}:
-                  </span>
-                  <div>
-                    <Badge variant="blue">
-                      {encounter.hospitalization?.re_admission
-                        ? t("re_admission")
-                        : t("new_admission")}
-                    </Badge>
-                  </div>
-                </div>
-              </div>
-            </div>
-            <Button
-              variant="outline"
-              className="hidden @sm:flex flex-row w-full text-gray-950"
-              asChild
-            >
-              <Link
-                href={`/facility/${encounter.facility.id}/patient/${encounter.patient.id}/encounter/${encounter.id}/questionnaire/encounter`}
-              >
-                <SquarePen className="size-3 text-gray-950" strokeWidth={1.5} />
-                <span className="font-semibold">{t("update_details")}</span>
-              </Link>
-            </Button>
-          </div>
-
-          <Separator className="sm:hidden" />
-
-          <div className="flex flex-row -mt-2 sm:mt-0 gap-5">
-            <div className="flex flex-col gap-2">
-              <div>
-                <span className="text-sm font-medium text-gray-700">
-                  {t("start_date")}:
-                </span>
-                <div className="text-sm text-gray-950 font-semibold">
-                  {encounter.period.start ? (
-                    <>
-                      {format(encounter.period.start, "dd MMM yyyy")},{" "}
-                      <span className="text-gray-600">
-                        {format(encounter.period.start, "hh:mma")}
-                      </span>
-                    </>
-                  ) : (
-                    <span>--</span>
-                  )}
-                </div>
-              </div>
-              <div>
-                <span className="text-sm font-medium text-gray-700">
-                  {t("account")}:
-                </span>
-                <div className="text-sm text-gray-950 font-semibold">
-                  {account?.results[0]?.name || "--"}
-                </div>
-              </div>
-            </div>
-
-            <div className="flex flex-col gap-2">
-              <div>
-                <span className="text-sm font-medium text-gray-700">
-                  {t("end_date")}:
-                </span>
-                <div className="text-sm text-gray-950 font-semibold">
-                  {encounter.period.end ? (
-                    <>
-                      {format(encounter.period.end, "dd MMM yyyy")},{" "}
-                      <span className="text-gray-600">
-                        {format(encounter.period.end, "hh:mma")}
-                      </span>
-                    </>
-                  ) : (
-                    <span>--({t("ongoing")})</span>
-                  )}
-                </div>
-              </div>
-
-              <div className="flex flex-col gap-2">
-                <div className="text-sm text-gray-950 font-semibold flex flex-wrap gap-2">
-                  {patient?.instance_identifiers?.map((identifier) => (
-                    <div
-                      key={identifier.config.id}
-                      className="flex md:flex-col gap-0.5 items-center md:items-start"
-                    >
-                      <span className="text-xs text-gray-600 w-32 md:w-auto">
-                        {identifier.config.config.display}:{" "}
-                      </span>
-                      <span className="text-sm font-semibold">
-                        {identifier.value}
-                      </span>
-                    </div>
-                  ))}
-                </div>
-              </div>
-
-              <div>
-                <span className="text-sm font-medium text-gray-700">
-                  {t("care_team")}:
-                </span>
-                <div className="flex items-center gap-2">
-                  <div className="flex -space-x-2">
-                    {encounter.care_team.length > 0 ? (
-                      encounter.care_team
-                        .slice(0, 3)
-                        .map((member) => (
-                          <Avatar
-                            key={member.member.id}
-                            name={member.member.first_name}
-                            imageUrl={member.member.profile_picture_url}
-                            className="size-8 rounded-full"
-                          />
-                        ))
-                    ) : (
-                      <span>--</span>
-                    )}
-                  </div>
-                  {encounter.care_team.length > 3 && (
-                    <span className="text-sm font-medium text-gray-600">
-                      +{encounter.care_team.length - 3}
-                    </span>
-                  )}
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-        <Button
-          variant="outline"
-          className="sm:hidden w-full text-gray-950"
-          asChild
-        >
-          <Link
-            href={`/facility/${encounter.facility.id}/patient/${encounter.patient.id}/encounter/${encounter.id}/questionnaire/encounter`}
-          >
-            <SquarePen className="size-3 text-gray-950" strokeWidth={1.5} />
-            <span className="font-semibold">{t("update_details")}</span>
-          </Link>
-        </Button>
-      </div>
-      <div className="@xs:hidden xl:flex flex-col gap-4">
+      <SummaryPanelEncounterDetails />
+      <div className="hidden xl:flex flex-col gap-4">
         <EncounterDetails />
         <EncounterTags />
         {canEdit && <Questionnaires />}
